fixed a number of const violations, missing, duplicate or wrong declarations etc.

+++ position/pieceid.h
@@ -11,12 +11,12 @@ enum
   MaxPieceId = 63
 };
 
-typedef unsigned int        PieceIdType;
+typedef unsigned long       PieceIdType;
 
 #define PieceIdOffset       (nr_piece_flags+1)
-#define PieSpMask           ((1<<PieceIdOffset)-1)
-#define PieceIdWidth        7
-#define PieceIdMask         (((1<<(PieceIdOffset+PieceIdWidth))-1)^((1<<PieceIdOffset)-1))
+#define PieSpMask           ((1u<<PieceIdOffset)-1)
+#define PieceIdWidth        7u
+#define PieceIdMask         (((1u<<(PieceIdOffset+PieceIdWidth))-1)^((1u<<PieceIdOffset)-1))
 #define SetPieceId(spec,id) ((spec) = ((id)<<PieceIdOffset) | ((spec)&PieSpMask))
 #define GetPieceId(spec)    ((spec) >> PieceIdOffset)
 #define ClearPieceId(spec)  SetPieceId(spec,NullPieceId)
